+++ src/atoms/Radio/Radio.stories.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { text, boolean, select } from '@storybook/addon-knobs';
 import { action } from '@storybook/addon-actions';
-import Radio from './RadioButton';
+import Radio from './Radio';
 import Flex from '../Flex';
 import View from '../View';
